+++ packages/protocol-devtools/src/uln302/schema.ts
@@ -10,6 +10,7 @@ export const Uln302ExecutorConfigSchema = z.object({
 export const Uln302UlnConfigSchema = z.object({
     confirmations: UIntBigIntSchema,
     requiredDVNs: z.array(AddressSchema),
+    requiredDVNCount: UIntNumberSchema,
     optionalDVNs: z.array(AddressSchema),
     optionalDVNThreshold: UIntNumberSchema,
 }) satisfies z.ZodSchema<Uln302UlnConfig, z.ZodTypeDef, unknown>
@@ -17,6 +18,7 @@ export const Uln302UlnConfigSchema = z.object({
 export const Uln302UlnUserConfigSchema = z.object({
     confirmations: UIntBigIntSchema.optional(),
     requiredDVNs: z.array(AddressSchema),
+    requiredDVNCount: UIntNumberSchema.optional(),
     optionalDVNs: z.array(AddressSchema).optional(),
     optionalDVNThreshold: UIntNumberSchema.optional(),
 }) satisfies z.ZodSchema<Uln302UlnUserConfig, z.ZodTypeDef, unknown>